refactor: move test files to tests directory
